add support for file generation with table list
You can now specify a list a tables which rule file
can be automatically created by rake.

Simply put the list in config.yml. No list means all
the tables.

With this feature, you can also tell the order of the
rules, just put the list in the wanted order.