fix false append mode
When a rule was using as new table the same table than a rule in append
mode, the converter thought it was in append mode too.

Many features was added in the last days. The code is ugly and it begins
to have some dirty hacks, like this one and precedent commits.

So for now, I'll stop adding features and I'll refactor the whole thing
( supposing no more bug fixes have to be done ).